Create VendorFundingForFreshAndPrimeNow
This query was used in our monthly business review to ensure that we were getting the expected vendor funding rates per sale of items that were sponsored by the vendor. When originally created, this was the first time we had this data available to either the Fresh or Prime Now business.